move the security allocator to prevent heart failure
